fix+perf: C chain — replace compact fp16/delta tail with exact-value/abspos tail
The compact 32-bit tail (fp16 << 16 | delta16) had a correctness bug:
when a new position overwrites tail[pos & WINDOW_MASK], the stale 16-bit
delta is reinterpreted relative to the CURRENT chain walk position, yielding
a wrong candidate address. This caused the chain to short-circuit after ~2
steps on inputs where WINDOW_SIZE positions share the same slot index —
i.e., every real-world block. Effect: chain=8 produced the same output as
chain=2 on all inputs.

Fix: use the Java-matching 64-bit tail layout:
  bits[63:32] = 4-byte src value at predecessor (rejection filter, avoids
                cold src[sv] load for non-matching candidates)
  bits[31:0]  = predecessor ABSOLUTE position (negative = NIL/end-of-chain)

With absolute positions, a stale slot is detected by the existing window
check (sv <= limit or sv < 0) and the chain walk terminates cleanly.

Result on large_test.bin (Mac M4 Pro, chain=8):
  Before: 280 MB/s, 2.60x ratio (broken — same as chain=2)
  After:  315 MB/s, 2.63x ratio (correct — matches Java output byte-for-byte)

C and Java now produce identical compressed output at every chain depth.
The tail table grows from 256 KiB to 512 KiB; total stream state is 768 KiB.

+++ src/main/native/lz4.h
@@ -31,16 +31,21 @@
 #define LZ4_HTAB_FAST_BYTES (LZ4_HASH_SIZE_FAST * sizeof(uint64_t))
 
 /*
- * Compact chain tail: uint32_t[WINDOW_SIZE] = 256 KiB (same size as int[]).
- * Each slot tail[pos & MASK] describes the link FROM pos TO its predecessor:
- *   bits[31:16] = 16-bit fingerprint of the predecessor (rejection filter)
- *   bits[15:0]  = 16-bit backward delta = pos - predecessor  (0 = no predecessor)
- * The fingerprint is (pos4 * 0x9E37u) >> 16 where pos4 is the 4-byte prefix of
- * the predecessor.  A mismatching fingerprint avoids loading src[predecessor].
+ * Chain tail: uint64_t[WINDOW_SIZE] = 512 KiB.
+ * Each slot tail[pos & MASK] stores the link FROM pos TO its predecessor:
+ *   bits[63:32] = 4-byte src value at the predecessor position (rejection filter)
+ *   bits[31:0]  = predecessor absolute position (NIL = 0x80808080 sentinel)
+ * Storing the absolute position (not a relative delta) avoids stale-slot
+ * corruption: when tail[pos & MASK] is overwritten by a later position that
+ * maps to the same slot, the stored absolute position fails the window check
+ * (sv <= limit) and the chain walk terminates cleanly.
+ * Storing the predecessor's 4-byte value avoids a cold src[sv] load on rejection.
+ * NIL sentinel: head[] is initialised to -1 (all 0xFF bytes), and an entry with
+ * prevPos < 0 (high bit set) is treated as end-of-chain.
  */
 typedef struct {
     int      head[LZ4_HASH_SIZE];
-    uint32_t tail[WINDOW_SIZE];
+    uint64_t tail[WINDOW_SIZE];
 } lz4_stream_t;
 
 /* Compress src_len bytes from src into dst.
